delete modal service and population of other user details on view modal

+++ src/app/accounts/admin/components/delete-modal/delete-modal.service.ts
@@ -1,4 +1,5 @@
 import { DOCUMENT } from '@angular/common';
+import { Observable } from 'rxjs';
 import {
   ComponentRef,
   Inject,
@@ -8,7 +9,9 @@ import {
   ViewContainerRef,
 } from '@angular/core';
 import { DeleteModalComponent } from './delete-modal.component';
-import { User } from '../../../../shared/types/types';
+import { User, GenericResponse } from '../../../../shared/types/types';
+import { UsersService } from '../../services/users.service';
+import { tap, finalize, EMPTY } from 'rxjs';
 
 @Injectable({
   providedIn: 'root',
@@ -16,7 +19,8 @@ import { User } from '../../../../shared/types/types';
 export class DeleteModalService {
   constructor(
     private injector: Injector,
-    @Inject(DOCUMENT) private document: Document
+    @Inject(DOCUMENT) private document: Document,
+    private usersService: UsersService
   ) {}
 
   open(viewContainerRef: ViewContainerRef, options?: { user?: User }) {
@@ -30,7 +34,7 @@ export class DeleteModalService {
       modalComponentRef.instance.user = options.user;
     }
     modalComponentRef.instance.deleteConfirmedEvent.subscribe(() =>
-      this.deleteConfirmed(modalComponentRef)
+      this.deleteConfirmed(modalComponentRef, options?.user?.email)
     );
     modalComponentRef.instance.cancelEvent.subscribe(() =>
       this.cancelModal(modalComponentRef)
@@ -40,8 +44,23 @@ export class DeleteModalService {
     return modalComponentRef;
   }
 
-  deleteConfirmed(modalComponentRef: ComponentRef<DeleteModalComponent>) {
-    this.destroyModal(modalComponentRef);
+  private deleteConfirmed(
+    modalComponentRef: ComponentRef<DeleteModalComponent>,
+    userId: string | undefined
+  ): Observable<GenericResponse> {
+    if (userId !== undefined) {
+      return this.usersService.deleteUser(userId).pipe(
+        tap(response => {
+          console.log(response);
+        }),
+        finalize(() => {
+          this.destroyModal(modalComponentRef);
+        })
+      );
+    } else {
+      console.error('User ID is undefined');
+      return EMPTY;
+    }
   }
 
   cancelModal(modalComponentRef: ComponentRef<DeleteModalComponent>) {
